Attempt to install a few pkgs correctly
It seems we need to update the database before doing this, or something
of that matter, as we're not finding the correct packages when we just
run things normally. I also changed the python-zypp install to be a
`pkg.installed` instead of a `cmd.run` so we could do a refresh as well.